pyproject.toml
[project]
name = "smart-visual-toolkit"
version = "0.1.0"
description = "MCP server for generating images with Nano Banana Pro styles"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.20.0",
    "google-genai>=0.2.0",
    "httpx>=0.28.1",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = [
    "src/main.py",  # MCP 서버 진입점 (인프라 코드)
    "*/__init__.py",  # 패키지 초기화 파일
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
